+++ wayland-backend/src/protocol.rs
@@ -2,7 +2,7 @@
 
 use std::{
     ffi::{CStr, CString},
-    os::unix::io::AsRawFd,
+    os::unix::io::{AsRawFd, BorrowedFd, OwnedFd},
 };
 
 #[cfg(any(test, feature = "client_system", feature = "server_system"))]
@@ -56,9 +56,9 @@ impl ArgumentType {
 }
 
 /// Enum of possible argument of the protocol
-#[derive(Debug, Clone)]
+#[derive(Debug)]
 #[allow(clippy::box_collection)]
-pub enum Argument<Id, Fd> {
+pub enum OwnedArgument<Id> {
     /// An integer argument. Represented by a [`i32`].
     Int(i32),
     /// An unsigned integer argument. Represented by a [`u32`].
@@ -79,13 +79,75 @@ pub enum Argument<Id, Fd> {
     /// The value is boxed to reduce the stack size of Argument. The performance
     /// impact is negligible as `array` arguments are pretty rare in the protocol.
     Array(Box<Vec<u8>>),
-    /// A file descriptor argument. Represented by a [`RawFd`].
-    ///
-    /// [`RawFd`]: std::os::fd::RawFd
-    Fd(Fd),
+    /// A file descriptor argument. Represented by a [`OwnedFd`].
+    Fd(OwnedFd),
 }
 
-impl<Id, Fd> Argument<Id, Fd> {
+impl<Id: PartialEq> PartialEq for OwnedArgument<Id> {
+    fn eq(&self, other: &Self) -> bool {
+        match (self, other) {
+            (Self::Int(a), Self::Int(b)) => a == b,
+            (Self::Uint(a), Self::Uint(b)) => a == b,
+            (Self::Fixed(a), Self::Fixed(b)) => a == b,
+            (Self::Str(a), Self::Str(b)) => a == b,
+            (Self::Object(a), Self::Object(b)) => a == b,
+            (Self::NewId(a), Self::NewId(b)) => a == b,
+            (Self::Array(a), Self::Array(b)) => a == b,
+            // Owned file descriptors must be different
+            (Self::Fd(_), Self::Fd(_)) => false,
+            _ => false,
+        }
+    }
+}
+
+impl<Id: Clone> OwnedArgument<Id> {
+    #[cfg(test)]
+    pub(crate) fn as_argument(&self) -> Argument<'_, Id> {
+        use std::os::fd::AsFd;
+        match self {
+            Self::Int(val) => Argument::Int(*val),
+            Self::Uint(val) => Argument::Uint(*val),
+            Self::Fixed(val) => Argument::Fixed(*val),
+            Self::Str(val) => Argument::Str(val.clone()),
+            Self::Object(val) => Argument::Object(val.clone()),
+            Self::NewId(val) => Argument::NewId(val.clone()),
+            Self::Array(val) => Argument::Array(val.clone()),
+            Self::Fd(val) => Argument::Fd(val.as_fd()),
+        }
+    }
+}
+
+impl<Id: Eq> Eq for OwnedArgument<Id> {}
+
+/// Enum of possible argument of the protocol
+#[derive(Debug, Clone)]
+#[allow(clippy::box_collection)]
+pub enum Argument<'a, Id> {
+    /// An integer argument. Represented by a [`i32`].
+    Int(i32),
+    /// An unsigned integer argument. Represented by a [`u32`].
+    Uint(u32),
+    /// A signed fixed point number with 1/256 precision
+    Fixed(i32),
+    /// CString
+    ///
+    /// The value is boxed to reduce the stack size of Argument. The performance
+    /// impact is negligible as `string` arguments are pretty rare in the protocol.
+    Str(Option<Box<CString>>),
+    /// Id of a wayland object
+    Object(Id),
+    /// Id of a newly created wayland object
+    NewId(Id),
+    /// `Vec<u8>`
+    ///
+    /// The value is boxed to reduce the stack size of Argument. The performance
+    /// impact is negligible as `array` arguments are pretty rare in the protocol.
+    Array(Box<Vec<u8>>),
+    /// A file descriptor argument. Represented by a [`BorrowedFd`].
+    Fd(BorrowedFd<'a>),
+}
+
+impl<'a, Id> Argument<'a, Id> {
     /// Retrieve the type of a given argument instance
     pub fn get_type(&self) -> ArgumentType {
         match *self {
@@ -99,22 +161,9 @@ impl<Id, Fd> Argument<Id, Fd> {
             Self::Fd(_) => ArgumentType::Fd,
         }
     }
-
-    fn map_fd<T>(self, f: &mut impl FnMut(Fd) -> T) -> Argument<Id, T> {
-        match self {
-            Self::Int(val) => Argument::Int(val),
-            Self::Uint(val) => Argument::Uint(val),
-            Self::Fixed(val) => Argument::Fixed(val),
-            Self::Str(val) => Argument::Str(val),
-            Self::Object(val) => Argument::Object(val),
-            Self::NewId(val) => Argument::NewId(val),
-            Self::Array(val) => Argument::Array(val),
-            Self::Fd(val) => Argument::Fd(f(val)),
-        }
-    }
 }
 
-impl<Id: PartialEq, Fd: AsRawFd> PartialEq for Argument<Id, Fd> {
+impl<'a, Id: PartialEq> PartialEq for Argument<'a, Id> {
     fn eq(&self, other: &Self) -> bool {
         match (self, other) {
             (Self::Int(a), Self::Int(b)) => a == b,
@@ -130,9 +179,25 @@ impl<Id: PartialEq, Fd: AsRawFd> PartialEq for Argument<Id, Fd> {
     }
 }
 
-impl<Id: Eq, Fd: AsRawFd> Eq for Argument<Id, Fd> {}
+impl<'a, Id: Eq> Eq for Argument<'a, Id> {}
 
-impl<Id: std::fmt::Display, Fd: AsRawFd> std::fmt::Display for Argument<Id, Fd> {
+impl<'a, Id: std::fmt::Display> std::fmt::Display for Argument<'a, Id> {
+    #[cfg_attr(unstable_coverage, coverage(off))]
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            Self::Int(value) => write!(f, "{value}"),
+            Self::Uint(value) => write!(f, "{value}"),
+            Self::Fixed(value) => write!(f, "{:.4}", *value as f64 / 256.0),
+            Self::Str(value) => write!(f, "{value:?}"),
+            Self::Object(value) => write!(f, "{value}"),
+            Self::NewId(value) => write!(f, "{value}"),
+            Self::Array(value) => write!(f, "{value:?}"),
+            Self::Fd(value) => write!(f, "{}", value.as_raw_fd()),
+        }
+    }
+}
+
+impl<Id: std::fmt::Display> std::fmt::Display for OwnedArgument<Id> {
     #[cfg_attr(unstable_coverage, coverage(off))]
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         match self {
@@ -308,34 +373,39 @@ pub struct ProtocolError {
 pub const INLINE_ARGS: usize = 4;
 
 /// Represents a message that has been sent from some object.
-#[derive(Clone, Debug)]
-pub struct Message<Id, Fd> {
+#[derive(Debug, PartialEq, Eq)]
+pub struct OwnedMessage<Id> {
     /// The id of the object that sent the message.
     pub sender_id: Id,
     /// The opcode of the message.
     pub opcode: u16,
     /// The arguments of the message.
-    pub args: smallvec::SmallVec<[Argument<Id, Fd>; INLINE_ARGS]>,
+    pub args: smallvec::SmallVec<[OwnedArgument<Id>; INLINE_ARGS]>,
 }
 
-impl<Id, Fd> Message<Id, Fd> {
-    /// Map some closure on all Fd contained in this message, to change the Fd generic parameter.
-    pub fn map_fd<T>(self, mut f: impl FnMut(Fd) -> T) -> Message<Id, T> {
+/*
+impl<Id: Clone> OwnedMessage<Id> {
+    #[cfg(test)]
+    pub(crate) fn as_message(&self) -> Message<'_, Id> {
         Message {
-            sender_id: self.sender_id,
+            sender_id: self.sender_id.clone(),
             opcode: self.opcode,
-            args: self.args.into_iter().map(move |arg| arg.map_fd(&mut f)).collect(),
+            args: self.args.iter().map(|arg| arg.as_argument()).collect(),
         }
     }
 }
+*/
 
-impl<Id: PartialEq, Fd: AsRawFd> PartialEq for Message<Id, Fd> {
-    fn eq(&self, other: &Self) -> bool {
-        self.sender_id == other.sender_id && self.opcode == other.opcode && self.args == other.args
-    }
+/// Represents a message that has been sent from some object.
+#[derive(Clone, Debug, PartialEq, Eq)]
+pub struct Message<'a, Id> {
+    /// The id of the object that sent the message.
+    pub sender_id: Id,
+    /// The opcode of the message.
+    pub opcode: u16,
+    /// The arguments of the message.
+    pub args: smallvec::SmallVec<[Argument<'a, Id>; INLINE_ARGS]>,
 }
-
-impl<Id: Eq, Fd: AsRawFd> Eq for Message<Id, Fd> {}
 
 impl std::error::Error for ProtocolError {}
 
@@ -365,10 +435,7 @@ pub fn same_interface(a: &'static Interface, b: &'static Interface) -> bool {
     std::ptr::eq(a, b) || a.name == b.name
 }
 
-pub(crate) fn check_for_signature<Id, Fd>(
-    signature: &[ArgumentType],
-    args: &[Argument<Id, Fd>],
-) -> bool {
+pub(crate) fn check_for_signature<Id>(signature: &[ArgumentType], args: &[Argument<Id>]) -> bool {
     if signature.len() != args.len() {
         return false;
     }